Add MapShed Job Foreign Key To Projects Table
In the case of Sub-basin GWLF-E modeling we send
GMS mapshed results for each HUC-12 that makes up
the AOI. Several dozen GMSs are often greater than
our 5MB nginx limit, so we will now allow the client
to instead send the job id of a successful mapshed
run. Add two new fields to the model to store the
successful job id: one for the unattenuated "tradiational"
mapshed results, and one for the sub-basin split mapshed
results job.