Allow Sending MapShed Job UUID To GWLF-E
In subbasin mode, the mapshed GMS payloads can grow to over
5MB (our nginx limit) when an AOI has more than about20 child
HUC-12s.

Allow sending the MapShed Job UUID to the GWLF-E endpoint so
we don't worry about exceeding the payload size.

Consequences:
* Using the job UUID to look up MapShed results puts pressure
on the jobs table which is not necessarily meant for long
term storage. We must now be careful if we ever need to empty
it, and we should consider other places to store the MapShed
results in the future.

* Modifications must now get burned in on the server-side